Repos update

+++ src/main/java/com/ruben/cementerio/repository/ServicioRepository.java
@@ -1,9 +1,10 @@
 package com.ruben.cementerio.repository;
 
-import com.ruben.cementerio.entity.Servicio;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import com.ruben.cementerio.entity.Servicio;
+
 @Repository
-public interface ServicioRepository extends JpaRepository<Servicio, Integer> {
+public interface ServicioRepository extends JpaRepository<Servicio, Long> {
 }
